finished renaming files and folders accordingly and placing relevant CSS classes in corresponding stylesheets
