+++ client/components/Validation/Validation.css
@@ -1,0 +1,5 @@
+.validation {
+  margin-top: 5%;
+  align-items: center;
+  text-align: center;
+}
